# perf(cli|main|report): lazy loads --init, reporter and caching code