--- a/src/main/index.mjs
+++ b/src/main/index.mjs
@@ -7,7 +7,6 @@ import extract from "../extract/index.mjs";
 import enrich from "../enrich/index.mjs";
 import cruiseResultSchema from "../schema/cruise-result.schema.mjs";
 import bus from "../utl/bus.mjs";
-import Cache from "../cache/cache.mjs";
 import {
   allExtensions,
   getAvailableTranspilers,
@@ -74,6 +73,8 @@ export async function cruise(
       c(2)
     );
 
+    const CacheModule = await import("../cache/cache.mjs");
+    const Cache = CacheModule.default;
     lCache = new Cache(lCruiseOptions.cache.strategy);
     const lCachedResults = lCache.read(lCruiseOptions.cache.folder);