--- a/src/report/dot/dot-custom.mjs
+++ b/src/report/dot/dot-custom.mjs
@@ -1,0 +1,3 @@
+import dotBase from "./index.mjs";
+
+export default dotBase("custom");