--- a/src/report/dot/dot-flat.mjs
+++ b/src/report/dot/dot-flat.mjs
@@ -1,0 +1,3 @@
+import dotBase from "./index.mjs";
+
+export default dotBase("flat");