--- a/src/report/dot/dot-folder.mjs
+++ b/src/report/dot/dot-folder.mjs
@@ -1,0 +1,3 @@
+import dotBase from "./index.mjs";
+
+export default dotBase("folder");