--- a/src/report/dot/dot-module.mjs
+++ b/src/report/dot/dot-module.mjs
@@ -1,0 +1,3 @@
+import dotBase from "./index.mjs";
+
+export default dotBase("module");