## Description, Motivation and Context

- lazily loads one reporter in the reporting stage
  ... as opposed to all of them on startup. Most reporters aren't very big or complicated, but all of them add up _and_ some of them still are relatively large compiled handlebars templates. 
  Lazy loading them typically shaves off ~40ms startup time on my (admittedly old) machine with the dependency-cruiser self-test.

- only loads the initConfig code when the --init option is specified
  ... we did this in version 12 as well as there it made a _huge_ impact (300ms) when we were still using `inquirer` for the interactive functionality. With the `prompts`  package it made a little less of a difference, but it was still measurable - and, as it turns out, it does under ESM as well.

- only loads the cache-ing code when the cache option is engaged
  of course this doesn't make a difference when running with the cache option (which one should), but if the cache option isn't used it also saves ~20ms

These tweaks seems to account for a net win of 80ms (real time) per run - see below for the results of the test runs.

> We also found that node 20 the whole run is 100ms faster than on node 19, probably because of the refactored ESM loading.

## How Has This Been Tested?

- [x] green ci
- [x] performance non-regression self test (`npm run test:load`) - see below

## Comparison

Both runs on a silent machine (2,6 GHz Quad-Core Intel Core i7) run with nodejs v20.0.0.

Note that the startup time for the dependency-cruiser self-run is quite unfavorable; it has vue (both vue2 and vue3), svelte, coffeescript, typescript, swc and babel installed - which makes all of these get loaded as well (removing the non-necessary brings down the startup time back from 1120ms to ~460ms). I hope for their DX in general in practice no-one will actually have such a mix in their repo ...

### Before (current `develop` branch)

```
$ hyperfine --warmup 3 --runs 30 "bin/dependency-cruise.mjs src bin test configs types tools --ignore-known --no-cache --no-progress"
Benchmark 1: bin/dependency-cruise.mjs src bin test configs types tools --ignore-known --no-cache --no-progress
  Time (mean ± σ):      3.165 s ±  0.023 s    [User: 4.809 s, System: 0.334 s]
  Range (min … max):    3.139 s …  3.255 s    30 runs
```

```
$ yarn depcruise --no-cache > /dev/null
 elapsed real          user        system         ∆ rss   ∆ heapTotal    ∆ heapUsed    ∆ external after step...
------------- ------------- ------------- ------------- ------------- ------------- ------------- ------------- 
     1.194 ms      1.335 ms        158 ms   +168.404 kB    +96.060 kB    +70.735 kB    +34.677 kB start of node process
        12 ms         13 ms          1 ms       +644 kB       +512 kB     +1.026 kB          0 kB parsing options
       190 ms        411 ms         12 ms    +10.936 kB    +12.316 kB    +11.368 kB       -157 kB parsing rule set
         1 ms          3 ms          0 ms       +160 kB       +256 kB        +93 kB          0 kB determining how to resolve
         0 ms          1 ms          0 ms         +4 kB          0 kB         +8 kB          0 kB reading files
        25 ms         45 ms          8 ms       +780 kB       +768 kB     +3.005 kB         +9 kB reading files: gathering initial sources
     1.212 ms      2.136 ms        111 ms    +46.684 kB    +45.056 kB    +44.679 kB       +364 kB reading files: visiting dependencies
         0 ms          1 ms          0 ms          0 kB          0 kB         +3 kB          0 kB analyzing
        17 ms         48 ms          2 ms     +2.340 kB     +2.304 kB     -6.122 kB          0 kB analyzing: cycles
        25 ms         71 ms          3 ms     +2.204 kB     +1.536 kB       +486 kB        -89 kB analyzing: dependents
         1 ms          1 ms          0 ms         +4 kB          0 kB        +46 kB          0 kB analyzing: orphans
       240 ms        339 ms          4 ms     +1.920 kB     +2.304 kB     +9.510 kB          0 kB analyzing: reachables
         0 ms          0 ms          0 ms          0 kB          0 kB         +3 kB          0 kB analyzing: module metrics
         0 ms          0 ms          0 ms         +4 kB          0 kB         +3 kB          0 kB analyzing: add focus (if any)
        79 ms        191 ms          2 ms       +580 kB       +256 kB     -2.948 kB          0 kB analyzing: validations
         4 ms          5 ms          0 ms        +72 kB          0 kB       +799 kB          0 kB analyzing: comparing against known errors
         4 ms          4 ms          0 ms        +16 kB          0 kB       +603 kB          0 kB reporting
         1 ms          1 ms          0 ms         +8 kB          0 kB         +4 kB          0 kB really done
------------- ------------- ------------- ------------- ------------- ------------- ------------- ------------- 
     3.006 ms      4.603 ms        302 ms   +234.760 kB   +161.368 kB   +133.302 kB    +34.804 kB 
```

### After (`refactor/lazy-load-reporters` branch)

```
$ hyperfine --warmup 3 --runs 30 "bin/dependency-cruise.mjs src bin test configs types tools --ignore-known --no-cache --no-progress"
Benchmark 1: bin/dependency-cruise.mjs src bin test configs types tools --ignore-known --no-cache --no-progress
  Time (mean ± σ):      3.062 s ±  0.019 s    [User: 4.687 s, System: 0.315 s]
  Range (min … max):    3.038 s …  3.123 s    30 runs
```

```
$ yarn depcruise --no-cache > /dev/null
 elapsed real          user        system         ∆ rss   ∆ heapTotal    ∆ heapUsed    ∆ external after step...
------------- ------------- ------------- ------------- ------------- ------------- ------------- ------------- 
     1.123 ms      1.270 ms        141 ms   +171.936 kB    +94.524 kB    +69.174 kB    +34.657 kB start of node process
        12 ms         14 ms          1 ms       +788 kB       +768 kB     +1.283 kB          0 kB parsing options
       193 ms        390 ms         11 ms    +11.776 kB    +12.060 kB    +11.534 kB       -158 kB parsing rule set
         0 ms          0 ms          0 ms          0 kB          0 kB        +28 kB          0 kB determining how to resolve
         0 ms          3 ms          0 ms        +12 kB          0 kB         +8 kB          0 kB reading files
        26 ms         29 ms          8 ms       +388 kB     +1.024 kB     +3.015 kB         +9 kB reading files: gathering initial sources
     1.209 ms      2.117 ms        106 ms    +42.064 kB    +44.800 kB    +44.737 kB       +364 kB reading files: visiting dependencies
         0 ms          0 ms          0 ms          0 kB          0 kB         +3 kB          0 kB analyzing
        15 ms         45 ms          2 ms     +2.344 kB     +2.560 kB     -6.992 kB          0 kB analyzing: cycles
        25 ms         79 ms          3 ms     +2.616 kB     +1.536 kB       +792 kB        -89 kB analyzing: dependents
         1 ms          1 ms          0 ms         +4 kB          0 kB        +47 kB          0 kB analyzing: orphans
       233 ms        343 ms          5 ms     +2.344 kB     +2.048 kB     +1.389 kB          0 kB analyzing: reachables
         0 ms          0 ms          0 ms          0 kB          0 kB         +3 kB          0 kB analyzing: module metrics
         0 ms          0 ms          0 ms          0 kB          0 kB         +3 kB          0 kB analyzing: add focus (if any)
        82 ms        200 ms          3 ms       +496 kB       +512 kB    +12.399 kB          0 kB analyzing: validations
         5 ms          5 ms          0 ms        +68 kB          0 kB       +800 kB          0 kB analyzing: comparing against known errors
        28 ms        101 ms          6 ms       +572 kB       +512 kB    -11.815 kB        +53 kB reporting
         1 ms          1 ms          0 ms         +4 kB          0 kB         +5 kB          0 kB really done
------------- ------------- ------------- ------------- ------------- ------------- ------------- ------------- 
     2.953 ms      4.599 ms        286 ms   +235.412 kB   +160.344 kB   +126.411 kB    +34.837 kB 
```

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] Documentation only change
- [x] Refactor (non-breaking change which fixes an issue without changing functionality)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist

- [x] :book:

  - My change doesn't require a documentation update, or ...
  - it _does_ and I have updated it

- [x] :balance_scale:
  - The contribution will be subject to [The MIT license](https://github.com/sverweij/dependency-cruiser/blob/develop/LICENSE), and I'm OK with that.
  - The contribution is my own original work.
  - I am ok with the stuff in [**CONTRIBUTING.md**](https://github.com/sverweij/dependency-cruiser/blob/develop/.github/CONTRIBUTING.md).